+++ CHANGELOG.md
@@ -9,9 +9,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+* `title` and `description` can be set from `pretty` and `description` keys at root.
+
+
 ### Changed
 
-* Moved `name` and `description` from `anyOf` to outside of schema. `anyOf` is now usable as intended for json schema.
+* Moved `title` and `description` from `anyOf` to outside of schema. `anyOf` is now usable as intended for json schema.
+
 
 ### Removed