+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+* Added git remote repositories and directories as potential sources for config files in the cli.
 * `title` and `description` can be set from `pretty` and `description` keys at root.
 
 
@@ -18,6 +19,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 
 ### Removed
+
+* Removed `title` and `description` support from `anyOf`.
 
 
 ## [0.3.0] - 2024-06-04