+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.4.0] - 2024-06-24
+
 ### Added
 
 * Added git remote repositories and directories as potential sources for config files in the cli.
@@ -80,7 +82,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Added contributors.
 
 
-[unreleased]: https://github.com/instructions-d-installation/installation-instruction/compare/v0.3.0...HEAD
+[unreleased]: https://github.com/instructions-d-installation/installation-instruction/compare/v0.4.0...HEAD
+[0.4.0]: https://github.com/instructions-d-installation/installation-instruction/compare/v0.3.0...v0.4.0
 [0.3.0]: https://github.com/instructions-d-installation/installation-instruction/compare/v0.2.0...v0.3.0
 [0.2.0]: https://github.com/instructions-d-installation/installation-instruction/compare/v0.1.1...v0.2.0
 [0.1.1]: https://github.com/instructions-d-installation/installation-instruction/compare/v0.1.0...v0.1.1